Refacor list pages: Extract client components and move to UI package
