+++ packages/ui/index.tsx
@@ -1,3 +1,4 @@
 export * from './badge';
 export * from './page';
 export * from './localized-markdown';
+export * from './small-screen-sub-menu';
