+++ packages/ui/package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@colonial-collections/list-store": "*",
     "@floating-ui/react": "0.24.0",
+    "@headlessui/react": "1.7.14",
     "@heroicons/react": "2.0.17",
     "classnames": "2.3.2",
     "next": "13.4.1",
